==> 4_module-trait_cor.qmd <==
---
title: "4_module-traits_correlations"
author: "Flavio Gabrieli"
format: html
editor: visual
---

# Module-trait correlation

Resume session

```{r}
session_name <- "TT8_RNAseq"
```

Load environment

```{r}
source("src/4_env_setup.R")
load(paste0(session_name, "/out/3_networks.Rdata"))
load(paste0(session_name, "/out/1_prepared_data.Rdata"))
source(paste0(session_name, "/in/", session_name,"_parameters.R"))
```

## Continuous variables

```{r}
if(!is.null(data$traits)){
  CTC <- continuous_traits_correlations(data, net)
  print(CTC)
  pdf(file = paste0(session_name,"/plots/4_cont_traits_correlation.pdf"), 
  width = 8,          # Width in inches
  height = 6,         # Height in inches
  bg = "white")
  print(CTC)
  dev.off()
}
```

## Build binary contrast matrix for categorical variables testing

<https://peterlangfelder.com/2018/11/25/working-with-categorical-variables/>

1.  you must set the desired baseline for each variable to test

```{r}
#set the reference value as the first level in the order
data$metadata$mat.gen <- factor(data$metadata$mat.gen, levels = c("Col-0", "tt8"))
data$metadata$mat.ploidy <- factor(data$metadata$mat.ploidy, levels = c("2x", "4x"))
data$metadata$pat.ploidy <- factor(data$metadata$pat.ploidy, levels = c("2x", "4x"))
data$metadata$viability <- factor(data$metadata$viability, levels = c("viable","unviable"))

cat("Categorical variable which module association can be tested:",colnames(data$metadata), sep = "\n")
```

```{r}
# generate the binary contrasts matrix for the desired variables to test
binary.matrix <- binary_contrast_matrix(data = data$metadata, 
                                        variables = colnames(data$metadata),
                                        vsAll = TRUE)
```

```{r}
#Specifiy whether to exclude some of the above contrasts and/or specify the prferred order
binary.matrix.order <- c(1,3,5,7,9:12,13,15,18,21,22)
binary.matrix <- binary.matrix[, binary.matrix.order]
head(binary.matrix)
```

## Categorical variables (using Pearson correlation)

```{r}
corr.cat.var <- correlation_catgorical_var(network = net, binary_matrix = binary.matrix)
corr.cat.var.km <- correlation_catgorical_var(network = net_km, binary_matrix = binary.matrix)
```

```{r}
pdf(file = paste0(session_name,"/plots/EMhc-condition_corr.pdf"), 
  width = 10,          # Width in inches
  height = 16.18,         # Height in inches
  bg = "white")
plot(corr.cat.var$hm)
dev.off()

pdf(file = paste0(session_name,"/plots/EMkm-condition_corr.pdf"), 
  width = 10,          # Width in inches
  height = 16.18,         # Height in inches
  bg = "white")
plot(corr.cat.var.km$hm)
dev.off()
```

## Categorical variables (using linear model)

The Estimate for the binary predictor indicates the difference in the mean eigengene value between the two groups defined by the binary contrast.

```{r}
plot.type = "dotplot" #choose between dotplot or heatmap

pdf(file = paste0(session_name,"/plots/EMkm-condition_corr_lm_", plot.type,".pdf"), 
  width = 10,          # Width in inches
  height = 16.18,         # Height in inches
  bg = "white")
correlation_categorical_var_lm(network = net_km, binary_matrix = binary.matrix, plot_type = plot.type)
dev.off()

pdf(file = paste0(session_name,"/plots/EMhc-condition_corr_lm_", plot.type,".pdf"), 
  width = 10,          # Width in inches
  height = 16.18,         # Height in inches
  bg = "white")
correlation_categorical_var_lm(network = net, binary_matrix = binary.matrix, plot_type = plot.type)
dev.off()


```

```{r}
rm(list = ls())
```
